- Adaptado hal de PIC18F26K80 a la nueva API

+++ grbl/pic18f26K80/missing.c
@@ -1,12 +1,13 @@
 #include "missing.h"
 #include <math.h>
 
+
 int32_t lround( float x )
 {
 #if 0
   return (x<0.0) ? x-0.5 : x+0.5;
 #else
-  return round(x);
+  return (int32_t)round(x);
 #endif
 }
